src/main/java/com/carter/leetcode/Item141.java
package com.carter.leetcode;

import com.carter.ListNode;

public class Item141 {
    public boolean hasCycle(ListNode head) {
        ListNode fast = head;
        ListNode low = head;
        while (fast != null && fast.next != null) {
            low = low.next;
            fast = fast.next.next;
            if (low == fast) {
                return true;
            }
        }
        return false;
    }

    public boolean hasCycle1(ListNode head) {
        if (head == null) {
            return false;
        }
        ListNode slow = head;
        ListNode fast = head.next;
        while (fast != null && fast.next != null) {
            if (fast == slow) {
                return true;
            }
            fast = fast.next;
            if (fast == slow) {
                return true;
            }
            fast = fast.next;
            slow = slow.next;
        }
        return false;
    }
}
